Added functionality to change ADC properties
There is DC motor that need to be controlled with PWM because diffrent
device's supply voltage. So to do that we need to switch between ADC
configurations (one for drawing variables and second for measuring
supply voltage). In datasheet manufacturer wrote that second adc
conversion result after changing channel or ref voltage is viable. So
second Conversion value is stored.